+++ meta_package_manager/tests/conftest.py
@@ -40,6 +40,7 @@ MANAGER_IDS = frozenset(
     [
         "apm",
         "apt",
+        "apt-mint",
         "brew",
         "cask",
         "composer",