+++ changelog.rst
@@ -8,6 +8,7 @@ Changelog
 .. note:: This version is not yet released and is under active development.
 
 * [mpm] Add a global ``-d/--dry-run`` option.
+* [apt] Add dedicated ``apt_mint`` manager to handle the special case of ``apt`` on Linux Mint.
 * [xbar] Let xbar plugin check minimal mpm version requirement.
 * [mpm] Add beta ``windows-2022`` CI/CD build target.