+++ changelog.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-`4.4.0 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v4.3.0...develop>`_
+`4.4.0 (2021-09-27) <https://github.com/kdeldycke/meta-package-manager/compare/v4.3.0...develop>`_
 --------------------------------------------------------------------------------------------------
 
 .. note:: This version is not yet released and is under active development.